api-server-python/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.10.1
pyodbc==5.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
